+++ Civi/ActionProvider/Action/Contact/GetEmailAddress.php
@@ -47,7 +47,16 @@ class GetEmailAddress extends AbstractGetSingleAction {
       $existingEmail = civicrm_api3('Email', 'getsingle', $existingEmailAddressParams);
       $this->setOutputFromEntity($existingEmail, $output);
     } catch (\Exception $e) {
-      // Do nothing
+      if($this->configuration->getParameter('fallbacktoactive')){
+        unset($existingEmailAddressParams['location_type_id']);
+        $existingEmailAddressParams['is_active'] = 1 ;
+        try {
+          $existingEmail = civicrm_api3('Email', 'getsingle', $existingEmailAddressParams);
+          $this->setOutputFromEntity($existingEmail, $output);
+        } catch (\Exception $e){
+
+        }
+      }
     }
   }
 
@@ -61,7 +70,8 @@ class GetEmailAddress extends AbstractGetSingleAction {
     reset($locationTypes);
     $defaultLocationType = key($locationTypes);
     return new SpecificationBag(array(
-      new Specification('location_type_id', 'Integer', E::ts('Location type'), true, $defaultLocationType, null, $locationTypes, FALSE)
+      new Specification('location_type_id', 'Integer', E::ts('Location type'), true, $defaultLocationType, null, $locationTypes, FALSE),
+      new Specification('fallbacktoactive', 'Boolean', E::ts('Fallback to active'), true, 0)
     ));
   }